+++ package.json
@@ -19,6 +19,7 @@
   },
   "homepage": "https://github.com/ylabio/react-webinar-3#readme",
   "dependencies": {
+    "@bem-react/classname": "^1.6.0",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   },
