+++ src/app.js
@@ -1,6 +1,8 @@
 import React from 'react';
-import {createElement} from './utils.js';
-import './styles.css';
+import List from "./components/list";
+import Controls from "./components/controls";
+import Head from "./components/head";
+import PageLayout from "./components/page-layout";
 
 /**
  * Приложение
@@ -9,35 +11,12 @@ import './styles.css';
  */
 function App({store}) {
 
-  const list = store.getState().list;
-
   return (
-    <div className='App'>
-      <div className='App-head'>
-        <h1>Приложение на чистом JS</h1>
-      </div>
-      <div className='App-controls'>
-        <button onClick={() => store.addItem()}>Добавить</button>
-      </div>
-      <div className='App-center'>
-        <div className='List'>{
-          list.map(item =>
-            <div key={item.code} className='List-item'>
-              <div className={'Item' + (item.selected ? ' Item_selected' : '')}
-                   onClick={() => store.selectItem(item.code)}>
-                <div className='Item-code'>{item.code}</div>
-                <div className='Item-title'>{item.title}</div>
-                <div className='Item-actions'>
-                  <button onClick={() => store.deleteItem(item.code)}>
-                    Удалить
-                  </button>
-                </div>
-              </div>
-            </div>
-          )}
-        </div>
-      </div>
-    </div>
+    <PageLayout>
+      <Head title='Приложение на чистом JS'/>
+      <Controls store={store}/>
+      <List store={store}/>
+    </PageLayout>
   );
 }
