+++ src/components/controls/style.css
@@ -1,0 +1,5 @@
+.Controls {
+  display: flex;
+  justify-content: flex-end;
+  padding: 20px;
+}
