+++ src/components/head/style.css
@@ -1,0 +1,9 @@
+.Head {
+  height: 80px;
+  background: #f5f5f5;
+}
+
+.Head h1 {
+  margin: 0;
+  padding: 20px 20px 10px;
+}
