--- src/styles.css
@@ -1,61 +1,0 @@
-body {
-  margin: 0;
-  padding: 0;
-  background: #535353;
-  font-family: sans-serif;
-}
-
-.App {
-  max-width: 1024px;
-  min-height: 100vh;
-  margin: 0 auto;
-  background: #fff;
-  box-sizing: border-box;
-  box-shadow: 0 0 15px 5px #0000003b;
-}
-
-.App-head {
-  height: 80px;
-  background: #f5f5f5;
-}
-
-.App-head h1 {
-  margin: 0;
-  padding: 20px 20px 10px;
-}
-
-.App-controls {
-  display: flex;
-  justify-content: flex-end;
-  padding: 20px;
-}
-
-.List-item {
-  border-bottom: 1px dashed #ccc;
-}
-
-.Item {
-  display: flex;
-  justify-content: flex-start;
-  transition: background-color 1s;
-}
-
-.Item_selected {
-  background: #75aaf1;
-  color: #fff;
-}
-
-.Item-code {
-  padding: 20px 10px 20px 20px;
-}
-
-.Item-title {
-  width: 100%;
-  padding: 20px 10px;
-}
-
-.Item-actions {
-  display: flex;
-  justify-content: flex-end;
-  padding: 20px 20px 20px 10px;
-}
